mailbox: t7: fifo mailbox for dsp and ao [2/4]
PD#SWPL-36392

Problem:
-fifo mailbox no support t7

Solution:
-add t7 mailbox driver

Verify:
T7 pxp

+++ bl30/src_ao/demos/amlogic/driver/mailbox/mailbox-irq.h
@@ -6,7 +6,7 @@ void vSetMbIrqHandler(unsigned int xNum, vHandlerFunc vHandler, void *vArg, unsi
 void vMbSetIrqPriority(unsigned int xNum, unsigned int xPriority);
 void vEnableMbInterrupt(unsigned int xMask);
 void vDisableMbInterrupt(unsigned int xMask);
-void vClrMbInterrupt(unsigned int xMask);
-unsigned int xGetMbIrqStats(void);
+void vClrMbInterrupt(uint64_t xMask);
+uint64_t xGetMbIrqStats(void);
 #endif
